Add and modify comments for split_dist set to 0.0 in the pulse voices.
Use the split_dist variable as a flag for pan in the generate frames routine
for the pulse voices.  This allowed removal of the test for whether it
was zero for split beat calculation.  Make sure it is set to 0 in setup.

Correct gnerate to generate for split_dist set to 0.0 in the chronaural voices.



git-svn-id: https://svn.code.sf.net/p/discord/code/trunk@111 928b5ed2-4cd8-4ab4-8030-20f29eb2a652