Change three of the included scripts to reflect the new layout for phase in discord.  These are phase_test.discord, beginner.discord, and intermediate.discord.
While I was looking at them I changed the sequences as well.  In beginner I
made it start at a lower beat frequency and put in carrier frequency slide as the
beat decreases.

Also in beginner, I removed the vim commands to change sequences as they are
obsolete now.

In both beginner and intermediate, I changed the pulse voice to be quieter but
to use a higher frequency.  Made it change very little as it drops.  The phase
voice is now in the center instead of in one channel or other for both as
well.

In test add amp beat and split beat time sequences.

Change the compatibility version to 3.0.0 in all three.



git-svn-id: https://svn.code.sf.net/p/discord/code/trunk@135 928b5ed2-4cd8-4ab4-8030-20f29eb2a652