This change converts discord so that it can use multiple script files simultaneously.  Internal threading so to speak.  This does two things.  It keeps interruptions from happening at node changes in the sound stream for continuous sounds and it lets the same beats be easily used with different accompaniment.  It is still backward compatible with any files from version 3.0.0 or greater and using single files still works fine.  Some of the major changes:
Add a new chorus_voice struct to hold each of the sound streams from the
separate script files.

Break script file parsing into two separate stages, that for options and that
for sound sequences.

Reverse the order of script file option parsing so that the options from the
last file have precedence.

Change the way that both playing and writing to a file are done to make them
simpler and in order to allow multiple sound streams to play together.

Because the generate frames function always gets a buffer that is pre zeroed,
removed the buffer zero set from that function.



git-svn-id: https://svn.code.sf.net/p/discord/code/trunk@145 928b5ed2-4cd8-4ab4-8030-20f29eb2a652