Modify discord to work with the new versions of libsndfile and libsamplerate.
Add the ability to write ogg files to discord in order to take advantage of
the new capability in libsndfile-1.0.20.  Add a new option, --vbr_quality/-y
that allows the setting of vbr quality in the ogg, from 0.0 (~ 64 Kb/s) to 
1.0 (~ 500Kb/s).  Default to .5, about 160 Kb/s.  Quite obviously the setting
is non-linear.

Tweak the option parsing routines a little bit for the parts related to
writing files in order to put out more meaningful error messages.

Find and fix an error in the file writing (and playing) portion of discord.
Didn't have any effect until ogg was enabled.  At the end of a sequence, an
empty buffer was being sent to write (or play).  Ignored by everything else,
but the vorbis library chokes when this happens.  Won't happen anymore from
discord.



git-svn-id: https://svn.code.sf.net/p/discord/code/trunk@164 928b5ed2-4cd8-4ab4-8030-20f29eb2a652