Add logic so repeat voices tie together at sequence nodes.
It has been the case that repeat voices start a new repetition of the
sample each time a sequence node occurs.  This creates discontinuity
in the sound.  In order to end that, this change has each repeat voice
node check if it is playing the same sound buffer as the previous
repeat voice node when it starts, and if it is, continue from where
the last voice left off.



git-svn-id: https://svn.code.sf.net/p/discord/code/trunk@166 928b5ed2-4cd8-4ab4-8030-20f29eb2a652