Add logic so sample voices tie together at sequence nodes.
It has been the case that sample voices start a new sample each time a
sequence node occurs.  In order to end that, this change has each
sample voice node check if it is playing the same sound buffer as the
previous sample voice node when it starts, and if it is, continue from
where the last voice left off.  The nature of the sampling process
means that there will always be discontinuities (clicks) at the sample
beginning and end;  this doesn't fix that, just ensures that the
sample is continued.
                                


git-svn-id: https://svn.code.sf.net/p/discord/code/trunk@167 928b5ed2-4cd8-4ab4-8030-20f29eb2a652