Add logic so stoch voices tie together at sequence nodes.
It has been the case that stoch voices create new stoch
characteristics each time a sequence node occurs.  In order to end
that, this change has each stoch voice node check if it is playing the
same sound buffer as the previous stoch voice node when it starts, and
if it is, continue from where the last voice left off.  The next
random play will start when it would have and if there is already a
play in progress it will finish.  At this point it does *not* check
for change in other characteristics at a node change.  Those changes
will take effect at the end of the next random play.  I think the use
case of changing amplitudes and splits at node junctures is rare, and
so this implementation makes sense.



git-svn-id: https://svn.code.sf.net/p/discord/code/trunk@168 928b5ed2-4cd8-4ab4-8030-20f29eb2a652