Add logic so once voices tie together at sequence nodes.
It has been the case that once voices are truncated when a sequence
node occurs.  In order to end that, this change has each once voice
node check if it is playing the same sound buffer as the previous once
voice node when it starts, and if it is, continue from where the last
voice left off if it didn't finish.  As much as possible of the
previous voice continuation plays before this once plays, even if it
can't play all of it.  For once it is perfectly legitimate to *not*
check for change in other characteristics at a node change.  Those
changes will take effect with the next once play.



git-svn-id: https://svn.code.sf.net/p/discord/code/trunk@169 928b5ed2-4cd8-4ab4-8030-20f29eb2a652