Add logic so bell voices tie together at sequence nodes.
It has been the case that bell voices are truncated when a sequence
node occurs.  In order to end that, this change has each bell voice
node check if it is playing the same carrier frequency and decay
pattern as the previous bell voice node when it starts, and if it is,
continue from where the last voice left off if it didn't finish.  This
also removes logic that prevented bell tones from initiating if they
wouldn't finish by the end of a sequence.




git-svn-id: https://svn.code.sf.net/p/discord/code/trunk@170 928b5ed2-4cd8-4ab4-8030-20f29eb2a652