Add logic so noise voices tie together at sequence nodes.
It has been the case that noise voices are truncated when a sequence
node occurs.  In order to end that, this change has each noise voice
node check if the carrier frequency range from the previous voice node
is contained in the carrier frequency range in the current voice node
when it starts, and if it is, continue from where the last voice left
off if it didn't finish.




git-svn-id: https://svn.code.sf.net/p/discord/code/trunk@171 928b5ed2-4cd8-4ab4-8030-20f29eb2a652