Change fade time to 10 ms and add fade in to noise voice.
The fade time has been 1 ms.  I noticed this was too abrupt as it
was causing noticeable artifacts in noise output.  And the sudden
start to a noise was also causing artifacts.  So the fade time was
changed to 10 ms and a fade in was added to the noise voice.



git-svn-id: https://svn.code.sf.net/p/discord/code/trunk@172 928b5ed2-4cd8-4ab4-8030-20f29eb2a652