First version of a spin voice added.
Given a sound sample, the spin voice will use phase, amplitude, and
split changes to create the impression that the sound is moving around
the listener at a requested spin rate.  This first version works, but
isn't entirely convincing.  It needs more spectral cues to get the
brain really thinking it is spinning.



git-svn-id: https://svn.code.sf.net/p/discord/code/trunk@180 928b5ed2-4cd8-4ab4-8030-20f29eb2a652