Add some buffer overflow protection to string processing.
While somewhat careful, discord was not paying much attention to
buffer overflow issues.  This commit adds a couple of functions, one
to allocate strings from memory, another to concatenate safely, to
help that.  In addition to allocating strings from memory, more
attention is now paid to freeing memory, though there are still
allocations only cleaned up at the end of running.  These changes work
with everything I've tried it with so far, but anytime pointers and
memory are messed with there is a higher potential for segmentation
faults / bugs.
                                 



git-svn-id: https://svn.code.sf.net/p/discord/code/trunk@181 928b5ed2-4cd8-4ab4-8030-20f29eb2a652