Fix long option handling from script files to be standard.
Long option handling was allowing single leading dashes when reading
options from script files.  This corrects that, and makes option
handling more robust in general.  It also adds vbr_quality to the help
message.  And it changes the return from parse_discordrc to be 0 on
success, 1 on failure, making it more standard.  In other words, some
cleanup maintenance.




git-svn-id: https://svn.code.sf.net/p/discord/code/trunk@182 928b5ed2-4cd8-4ab4-8030-20f29eb2a652