Add more structure to support an eventual spin voice.
Add a conversion to mono for the spin voice sample, and a second
filtered buffer with some primitive averaging of the samples.
Simplify the code in the generate frames routine because of the mono
input.  Not sure how far this will go.  Converting an arbitrary input
to have spin effect is very compute intensive, and must be done on
each sample that is played.  I've also found while listening to some
excellent samples on the web, that they still don't create the
impression of passing in front of the body even with all the
processing.  Will continue working on this, but it seems long term.
Will need to get some heavy duty algorithms involved, I think.  There
is also lots of room for tuning in the existing algorithm.



git-svn-id: https://svn.code.sf.net/p/discord/code/trunk@187 928b5ed2-4cd8-4ab4-8030-20f29eb2a652