Add support for true saw wave and reverse.
The wave form that I thought was a saw wave is not really a saw.  I
have reclassified it as half-saw and added a true saw and reverse saw.
Again, compiled cleanly, and the program runs fine with default set to
sin, but no testing of these new wave forms was performed.



git-svn-id: https://svn.code.sf.net/p/discord/code/trunk@189 928b5ed2-4cd8-4ab4-8030-20f29eb2a652