Add two new long options, --auto_volume and --max_volume, to the program. auto_volume adjusts the volume at the beginning of a segment to be the volume that is set in proportion to the volumes of the voices.  max_volume ensures that the volume at the start of a segment never exceeds the given volume.  When both are specified, if auto_volume is greater than max_volume, then max_volume is set, otherwise it is auto_volume.  max_volume only comes into effect when the requested volume exceeds the threshold, auto_volume is active anytime the volume is different than the requested volume.
They work fine when the number of voices in every segment is the same.  But if the
number of voices changes in segments, and there are slides, the volumes might 
behave differently than you would expect.  So they are flawed.  But they probably
have their place so I am leaving them in.  Expect them to be fuzzy.

I've looked at correcting these flaws, but it isn't clear that it would be more correct 
than the currently implemented version.


git-svn-id: https://svn.code.sf.net/p/discord/code/trunk@192 928b5ed2-4cd8-4ab4-8030-20f29eb2a652